refactor: moved engine options to type files
